refactor: split commands/enhancement.py into _enhancement_helpers
Deduplicate _offer_trafilatura_if_missing by sharing from
_context_helpers. Extract enhancement-specific logic
(enhance_text, input resolution, format inference) into
_enhancement_helpers.py. Update test patch targets to match
new module locations.
